bug fixes to compile with gfortran

+++ src/differentiation/hessian_tgt/b2us_geo_diffv_diffv.F90
@@ -2243,7 +2243,7 @@ CONTAINS
     INTEGER :: result12
     INTEGER :: result13
     REAL(r8) :: result14
-    REAL(kind=r8) :: result15
+    INTEGER :: result15
     REAL(kind=r8) :: arg1
 !
     ncv = mpg%ncv
