+++ src/differentiation/hessian_tgt/b2us_map_diffv_diffv.F90
@@ -1704,11 +1704,11 @@ CONTAINS
     EXTERNAL IPGETI, XERRAB
     INTRINSIC MAXVAL
     EXTERNAL MAXVAL_DV
-    REAL(kind=r8) :: result1
+    INTEGER :: result1
     INTEGER :: nd
     INTEGER :: nbdirs
     REAL(kind=r8) :: MAXVAL_DV
-    REAL :: result10
+    INTEGER :: result10
 !wdk  This initialization routine precomputes a number of arrays
 !wdk  in the mapping that are not stored in the b2fgmtry file, but
 !wdk  can be derived directly from data in that file
@@ -2222,7 +2222,7 @@ CONTAINS
     LOGICAL :: duplicate, left_match_found, right_match_found
     EXTERNAL IPGETI, XERRAB
     INTRINSIC MAXVAL
-    REAL(kind=r8) :: result1
+    INTEGER :: result1
 !wdk  This initialization routine precomputes a number of arrays
 !wdk  in the mapping that are not stored in the b2fgmtry file, but
 !wdk  can be derived directly from data in that file
